requirements.txt
beautifulsoup4
requests
googletrans==4.0.0-rc1
nltk
selenium==4.8.0
pytest==8.3.4
pytest-xdist==3.6.1
browserstack-local
urllib3<2.0
python-dotenv
